// Web/RMA-Web-Shell/projects/clientcare/src/app/product-manager/models/product-option-summary.ts
export class ProductOptionSummary {

  optionCode: string;
  optionName: string;
  optionDescription: string;
  startDate: string;
  endDate: string;
  maxAdminFeePercentage: number;
  maxCommissionFeePercentage: number;
  maxBinderFeePercentage: number;
  optionStatus: string;
  scale: string;
}